fix/changed kanban to component
changed from page to component
added styling
