+++ src/App.js
@@ -4,7 +4,7 @@ import { BrowserRouter, Routes, Route } from "react-router-dom";
 import Landing from './pages/landing';
 import Login from './pages/login';
 import Signup from './pages/signup';
-import Kanban from './pages/kanban';
+
 
 function App() {
   return (
@@ -13,7 +13,7 @@ function App() {
         <Route exact path="/" element={<Landing />}></Route>
         <Route exact path="/login" element={<Login />} > </Route>
         <Route exact path="/signup" element={<Signup />} > </Route>
-        <Route exact path="/dashboard" element={<Kanban />} > </Route>
+        
       </Routes>
     </BrowserRouter>
   );
